+++ edition/voxel_tool_terrain.h
@@ -4,6 +4,8 @@
 #include "voxel_tool.h"
 
 class VoxelTerrain;
+class VoxelDataMap;
+class VoxelLibrary;
 
 class VoxelToolTerrain : public VoxelTool {
 	GDCLASS(VoxelToolTerrain, VoxelTool)
@@ -28,6 +30,13 @@ public:
 
 	void run_blocky_random_tick(
 			AABB voxel_area, int voxel_count, const Callable &callback, int block_batch_count) const;
+
+	// For easier unit testing (the regular one needs a terrain setup etc, harder to test atm)
+	// The `_static` suffix is because it otherwise conflicts with the non-static method when registering the class
+	static void VoxelToolTerrain::run_blocky_random_tick_static(VoxelDataMap &map, Box3i voxel_box,
+			const VoxelLibrary &lib, int voxel_count, int batch_count, void *callback_data,
+			bool (*callback)(void *, Vector3i, int64_t));
+
 	void for_each_voxel_metadata_in_area(AABB voxel_area, const Callable &callback);
 
 protected: